pyproject.toml
[build-system]
requires = ["uv-build"]
build-backend = "uv_build"

[project]
name = "python-ai-service"
version = "0.1.0"
description = "Python AI Service"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi",
    "langchain",
    "langchain-openai",
    "langgraph",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "pydantic",
    "uvicorn",
]

[dependency-groups]
dev = [
    "black",
    "pytest",
    "pytest-mock",
    "ruff",
]

[tool.ruff]
target-version = "py313"


[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--tb=short --import-mode=importlib"
